requirements.txt
crewai
langchain_openai
fastapi
uvicorn
python-multipart
pdf2docx
markitdown[docx]
python-dotenv
pdfplumber
litellm
